vere, u3: move events module from noun/ to vere/
