+++ pkg/urbit/vere/event_log.c
